[wip] validate and analyse some assumptions about subprojects
this was a quick hack using python and no inherent knowledge of
what owners are, except for trying to expand aliases, use at your
own risk, avert your eyes children, etc

warn on a few criteria:
- sigs that don't have subprojects
- subprojects that don't have any owners files present
- missing owners files
- missing common labels (should at least have sig/foo)
- missing common approvers (closest thing we have to approvers)

other ideas to consider:
- too many common approvers
- do we care about common reviewers
- sig chairs in too many subproject files
- are there owners files out there that should fall under subprojects

This dumps a lot of stuff back into sigs.yaml to quickly see things
at a glance. It does not attempt to make PR's back, and it does not
attempt to file issues. I think that's an easy next step for sig
labels (and maybe subproject labels down the line).